feat(integrations): publish evidence-bounded agent compatibility
